refactor: separate sync worker from ui state
