+++ KMReader/Features/Sync/Models/SyncProgress.swift
@@ -1,0 +1,14 @@
+//
+// SyncProgress.swift
+//
+//
+
+import Foundation
+
+nonisolated struct SyncProgress: Sendable {
+  let phase: SyncPhase
+  let phaseProgress: Double
+  let stage: SyncStage?
+}
+
+typealias SyncProgressHandler = @MainActor @Sendable (SyncProgress) -> Void
