+++ KMReader/Features/Sync/Models/SyncRequest.swift
@@ -1,0 +1,11 @@
+//
+// SyncRequest.swift
+//
+//
+
+import Foundation
+
+nonisolated struct SyncRequest: Sendable {
+  let instanceId: String
+  let forceFullSync: Bool
+}
