+++ KMReader/Features/Sync/Models/SyncResult.swift
@@ -1,0 +1,11 @@
+//
+// SyncResult.swift
+//
+//
+
+import Foundation
+
+nonisolated struct SyncResult: Sendable {
+  let hasFailures: Bool
+  let readingProgressSynced: Bool
+}
